cleaned up, catched a NullPointerException, added logger to the preprocessing utils class
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@332 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/preprocessing/map/BinaryFileWriter.java
@@ -234,8 +234,8 @@ public class BinaryFileWriter {
 
 				// *** TEMP ***//
 
-				minZoom = (short) ((short) zoom_level - 2);
-				maxZoom = (short) ((short) zoom_level + 2);
+				minZoom = (short) (zoom_level - 2);
+				maxZoom = (short) (zoom_level + 2);
 
 				// *** TEMP ***//
 
@@ -886,6 +886,7 @@ public class BinaryFileWriter {
 		}
 
 		try {
+
 			raf = new RandomAccessFile(targetFile, "rw");
 
 			BinaryFileWriter bfw = new BinaryFileWriter(propertiesFile);
